Generic smart Media Section component reusable for the admin back-office
