--- src/components/adminMedia/adminMediaRow.module.scss
@@ -1,23 +1,0 @@
-@import "../../index";
-
-.adminMediaRowContainer {
-
-  max-width: 1300px;
-  width: 100%;
-
-  .title {
-    @extend .flexRow;
-    @extend .flexAlignCenter;
-    text-transform: lowercase;
-    letter-spacing: 1px;
-    color: $color-bright-1;
-    margin: 0 10px 15px;
-
-    .imageContainer {
-      height: 30px;
-      width: 30px;
-      margin-right: 10px;
-    }
-  }
-
-}
