+++ src/components/mediaSection/mediaSection.module.scss
@@ -1,0 +1,52 @@
+@import "../../index";
+
+
+.mediaSectionContainer {
+  .mediaRowContainer {
+    width: 100%;
+
+    .mediaRowContainer {
+      padding-top: 20px;
+      width: 100%;
+      padding-bottom: 20px;
+      @extend .flexColumn;
+      @extend .flexAlignCenter;
+      transition: background-color 0.3s ease-in-out, max-height 0.3s ease-in-out;
+      overflow: hidden;
+
+      .activeMediaContainer {
+        opacity: 0;
+        transition: all 0.2s ease-in-out 0.2s;
+      }
+
+      &.mediaRowContainerActive {
+        background-color: $color-dark-2;
+
+        .activeMediaContainer {
+          opacity: 1;
+        }
+      }
+    }
+
+    .mediaRowWrapper {
+      @extend .flexColumn;
+      max-width: 1300px;
+      width: 100%;
+    }
+
+    .title {
+      @extend .flexRow;
+      @extend .flexAlignCenter;
+      text-transform: lowercase;
+      letter-spacing: 1px;
+      color: $color-bright-1;
+      margin: 0 10px 5px;
+
+      .imageContainer {
+        height: 30px;
+        width: 30px;
+        margin-right: 10px;
+      }
+    }
+  }
+}
