+++ src/pages/admin/admin.module.scss
@@ -43,3 +43,7 @@
   }
 
 }
+
+.adminPageContainer {
+  width: 100%;
+}
