+++ src/pages/game/GamePage.js
@@ -1,21 +1,17 @@
 import styles from "./gamePage.module.scss";
 import React from 'react';
 import {Helmet} from "react-helmet";
-import cx from "classnames";
 import {getGameById} from "../../endpoints/gamesEndpoint";
 import PageLayout from "../../layouts/PageLayout";
 import moment from "moment";
 import localization from 'moment/locale/fr';
 import {withRouter} from "react-router-dom";
 import {LoadingSpinner} from "../../components/loadingSpinner/loadingSpinner";
-import Carousel from "../../components/carousel/carousel";
-import ActiveMediaBox from "../../components/activeMediaBox/activeMediaBox";
 import {connect} from "react-redux";
 import {ACTIONS_MEDIAS} from "../../actions/mediaActions";
 import {MEDIA_TYPES} from "../../config/const";
-import {setGamesForMedia, toggleVerifyMedia} from "../../endpoints/mediasEndpoint";
 import {ACTIONS_GAMES} from "../../actions/gamesActions";
-import {findPos} from "../../utils";
+import MediaSection from "../../components/mediaSection/mediaSection";
 
 class GamePage extends React.Component {
 
@@ -25,10 +21,6 @@ class GamePage extends React.Component {
         this.state = {
             game: {}
         };
-
-        this._handleClickMedia = this._handleClickMedia.bind(this);
-        this._handleSaveGames = this._handleSaveGames.bind(this);
-        this._handleVerifyMedia = this._handleVerifyMedia.bind(this);
     }
 
     componentDidMount() {
@@ -55,105 +47,14 @@ class GamePage extends React.Component {
             })
     }
 
-    _handleClickMedia(episode, ref) {
-        this.props.dispatch({
-            type: ACTIONS_MEDIAS.SET_ACTIVE_MEDIA,
-            payload: {...episode}
-        });
-        setTimeout(() => {
-            window.scrollTo({top: findPos(ref.current.offsetParent.offsetParent.offsetParent), behavior: 'smooth'})
-        }, 200);
-    }
-
-    _handleSaveGames(games) {
-        return setGamesForMedia({
-            games: games.map((game) => {
-                const mappedGame = {...game};
-                MEDIA_TYPES.forEach((mediaType) => {
-                    mediaType.medias.forEach((media) => {
-                        mappedGame[media.dataLabel] = null;
-                    });
-                    if (isNaN(mappedGame.releaseDate))
-                        mappedGame.releaseDate = null;
-                });
-                return mappedGame;
-            }),
-            mediaType: this.props.mediaActive.media.type,
-            mediaId: this.props.mediaActive.media.id
+    sortEpisodesByMediaTypes(episodes) {
+        return MEDIA_TYPES.map((type) => {
+            return {
+                ...type,
+                type: type.dataLabel,
+                episodes: episodes.filter((episode) => episode.media.type === type.dataLabel)
+            }
         })
-            .then(() => {
-                const gameId = String(this.props.match.params.gameId);
-                getGameById(gameId)
-                    .then((game) => {
-                        this.setState({game: game, loading: false});
-                    });
-                return true;
-            })
-    }
-
-    _handleVerifyMedia() {
-        return toggleVerifyMedia({
-            verified: true,
-            mediaType: this.props.mediaActive.mediaType,
-            mediaId: this.props.mediaActive.media.id
-        })
-            .then(() => {
-                this.setState({
-                    medias: this.state.medias.map((media) => {
-                        if (media.id === this.props.mediaActive.media.id) {
-                            return {...media, isVerified: true};
-
-                        }
-                        return media;
-                    })
-                });
-                this.props.dispatch({
-                    type: ACTIONS_MEDIAS.SET_ACTIVE_MEDIA,
-                    payload: {
-                        mediaType: this.props.type.dataLabel,
-                        media: {...this.props.mediaActive.media, isVerified: true}
-                    }
-                });
-                return true;
-            })
-    }
-
-    renderActiveMedia(mediaActive) {
-        return <ActiveMediaBox media={mediaActive} onSaveGames={this._handleSaveGames}
-                               onVerifyMedia={this._handleVerifyMedia}
-                               hideDescription={mediaActive.media.type === "video"}/>
-    }
-
-    renderMediaTypeRow(mediaType) {
-        const {mediaActive} = this.props;
-        let mediaActiveType;
-        if (mediaActive && mediaActive.media) {
-            mediaActiveType = MEDIA_TYPES.find(medType => medType.dataLabel === mediaActive.media.type);
-        }
-        const medias = this.props.currentGame.episodes.filter(ep => ep.media.type === mediaType.dataLabel);
-        const activeItem = this.props.mediaActive;
-
-        if (medias && medias.length > 0) {
-            return <div key={mediaType.dataLabel} ref={mediaType.ref}
-                        className={cx(styles.mediaRowContainer, {[styles.mediaRowContainerActive]: mediaActive && mediaActive.media && mediaActiveType && mediaType.dataLabel === mediaActiveType.dataLabel})}>
-                <div className={styles.mediaRowWrapper}>
-                    <div className={styles.title}>
-                        <img className={styles.imageContainer} src={mediaType.logoMin} alt={mediaType.dataLabel}/>
-                        {mediaType.name}
-                    </div>
-                    <Carousel medias={medias}
-                              onClickItem={(media, ref) => {
-                                  this._handleClickMedia(media, ref)
-                              }}
-                              activeItem={activeItem}
-                              smallerCards={mediaType.dataLabel === "videos"}/>
-                    <div className={styles.activeMediaContainer}>
-                        {mediaActive && mediaActive.media && mediaType.dataLabel === mediaActiveType.dataLabel && this.renderActiveMedia(mediaActive)}
-                    </div>
-                </div>
-            </div>
-        }
-        return null;
     }
 
     render() {
@@ -177,14 +78,7 @@ class GamePage extends React.Component {
                             </div>
                         </div>
                     </div>
-
-                    <div className={styles.gameContentWrapper}>
-                        <div className={styles.gameContent}>
-                            {MEDIA_TYPES.map((mediaType) => {
-                                return this.renderMediaTypeRow(mediaType);
-                            })}
-                        </div>
-                    </div>
+                    <MediaSection medias={this.sortEpisodesByMediaTypes(currentGame.episodes)}/>
                 </div>}
         </PageLayout>
     }
